+++ src/sentry/integrations/vercel/__init__.py
@@ -1,5 +1,4 @@
 from .client import *  # noqa: F401,F403
-from .generic_webhook import *  # noqa: F401,F403
 from .integration import *  # noqa: F401,F403
 from .urls import *  # noqa: F401,F403
 from .webhook import *  # noqa: F401,F403
