+++ src/sentry/integrations/vercel/urls.py
@@ -2,21 +2,17 @@ from django.conf.urls import url
 
 from sentry.web.frontend.vercel_extension_configuration import VercelExtensionConfigurationView
 
-from .generic_webhook import VercelGenericWebhookEndpoint
 from .webhook import VercelWebhookEndpoint
 
 urlpatterns = [
-    url(r"^webhook/$", VercelWebhookEndpoint.as_view(), name="sentry-extensions-vercel-webhook"),
     url(
         r"^configure/$",
         VercelExtensionConfigurationView.as_view(),
         name="sentry-extensions-vercel-configure",
     ),
-    # XXX(meredith): This route has become our generic hook, in
-    # the future we'll need to update the route name to reflect that.
-    url(
-        r"^delete/$",
-        VercelGenericWebhookEndpoint.as_view(),
-        name="sentry-extensions-vercel-generic-webhook",
-    ),
+    # Since we've been endorsing using `/delete` as the endpoint for Self-Hosted, we need to
+    # keep it operational for existing integrations. This is purely aesthetic, as both routes
+    # will use the same webhook (Previously known as 'Generic Webhook' - See #26185)
+    url(r"^delete/$", VercelWebhookEndpoint.as_view(), name="sentry-extensions-vercel-delete"),
+    url(r"^webhook/$", VercelWebhookEndpoint.as_view(), name="sentry-extensions-vercel-webhook"),
 ]
